+++ task-parser/src/main/java/com/oppo/cloud/parser/service/job/parser/CommonTextParser.java
@@ -17,8 +17,8 @@
 package com.oppo.cloud.parser.service.job.parser;
 
 import com.oppo.cloud.common.util.textparser.*;
+import com.oppo.cloud.parser.domain.job.ParserParam;
 import com.oppo.cloud.parser.domain.reader.ReaderObject;
-import com.oppo.cloud.parser.service.job.oneclick.ParserListenerBus;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.BufferedReader;
@@ -29,6 +29,11 @@ public abstract class CommonTextParser extends IParser {
 
     private ReaderObject readerObject;
     private List<ParserAction> actions;
+
+    public CommonTextParser(ParserParam param, List<ParserAction> actions) {
+        super(param);
+        this.actions = actions;
+    }
 
     public Map<String, ParserAction> parse(ReaderObject readerObject, List<ParserAction> actions) throws Exception {
         this.readerObject = readerObject;
@@ -62,4 +67,7 @@ public abstract class CommonTextParser extends IParser {
         return headTextParser.getResults();
     }
 
+    public List<ParserAction> getActions() {
+        return actions;
+    }
 }
