+++ task-parser/src/main/java/com/oppo/cloud/parser/service/job/parser/IParser.java
@@ -16,13 +16,38 @@
 
 package com.oppo.cloud.parser.service.job.parser;
 
+import com.oppo.cloud.common.constant.ProgressState;
+import com.oppo.cloud.common.domain.oneclick.OneClickProgress;
+import com.oppo.cloud.common.domain.oneclick.ProgressInfo;
 import com.oppo.cloud.parser.domain.job.CommonResult;
+import com.oppo.cloud.parser.domain.job.ParserParam;
 import com.oppo.cloud.parser.service.job.oneclick.ParserListenerBus;
 
 public abstract class IParser extends ParserListenerBus {
 
+    protected ParserParam param;
+
+    public IParser(ParserParam param) {
+        this.param = param;
+    }
+
     public CommonResult run() {
         return null;
+    }
+
+    public void updateParserProgress(ProgressState state, Integer progress, Integer count) {
+        if (!this.param.getLogRecord().getIsOneClick()) {
+            return;
+        }
+        OneClickProgress oneClickProgress = new OneClickProgress();
+        oneClickProgress.setAppId(this.param.getApp().getAppId());
+        oneClickProgress.setLogType(this.param.getLogType());
+        ProgressInfo executorProgress = new ProgressInfo();
+        executorProgress.setCount(count);
+        executorProgress.setProgress(progress);
+        executorProgress.setState(state);
+        oneClickProgress.setProgressInfo(executorProgress);
+        super.update(oneClickProgress);
     }
 
 }
