+++ task-parser/src/main/java/com/oppo/cloud/parser/service/job/parser/MapReduceContainerLogParser.java
@@ -18,8 +18,6 @@ package com.oppo.cloud.parser.service.job.parser;
 
 import com.oppo.cloud.common.constant.ProgressState;
 import com.oppo.cloud.common.domain.job.LogPath;
-import com.oppo.cloud.common.domain.oneclick.OneClickProgress;
-import com.oppo.cloud.common.domain.oneclick.ProgressInfo;
 import com.oppo.cloud.common.util.textparser.ParserAction;
 import com.oppo.cloud.parser.domain.job.CommonResult;
 import com.oppo.cloud.parser.domain.job.ParserParam;
@@ -36,18 +34,9 @@ import java.util.Map;
 @Slf4j
 public class MapReduceContainerLogParser extends CommonTextParser {
 
-    private final ParserParam param;
-
-    private boolean isOneClick;
-
-    private List<ParserAction> actions;
-
     public MapReduceContainerLogParser(ParserParam param,
                                        List<ParserAction> actions) {
-
-        this.param = param;
-        this.actions = actions;
-        this.isOneClick = param.getLogRecord().getIsOneClick();
+        super(param, actions);
     }
 
     @Override
@@ -71,7 +60,7 @@ public class MapReduceContainerLogParser extends CommonTextParser {
             for (ReaderObject readerObject : readerObjects) {
                 Map<String, ParserAction> results;
                 try {
-                    results = super.parse(readerObject, actions);
+                    results = super.parse(readerObject, this.getActions());
                 } catch (Exception e) {
                     log.error("Exception:", e);
                     continue;
@@ -89,18 +78,4 @@ public class MapReduceContainerLogParser extends CommonTextParser {
         return commonResult;
     }
 
-    public void updateParserProgress(ProgressState state, Integer progress, Integer count) {
-        if (!this.isOneClick) {
-            return;
-        }
-        OneClickProgress oneClickProgress = new OneClickProgress();
-        oneClickProgress.setAppId(this.param.getApp().getAppId());
-        oneClickProgress.setLogType(this.param.getLogType());
-        ProgressInfo executorProgress = new ProgressInfo();
-        executorProgress.setCount(count);
-        executorProgress.setProgress(progress);
-        executorProgress.setState(state);
-        oneClickProgress.setProgressInfo(executorProgress);
-        super.update(oneClickProgress);
-    }
 }
